refactor(simd): Split expression.rs into focused submodules
Split the 1183-line simd/expression.rs file into a directory module
with focused submodules by concern:

- analysis.rs: Expression analysis for SIMD eligibility
  (can_use_simd_for_expression, is_simple_operand, extract_column_refs)

- null_handling.rs: NULL value detection for graceful fallback
  (has_null_values function)

- evaluation.rs: Core SIMD evaluation logic and buffer management
  (eval_expression_batch_simd, apply_simd_operation, NumericValue,
  NumericBuffer, etc.)

- mod.rs: Public API, constants (SIMD_THRESHOLD, MAX_RECURSION_DEPTH),
  and module documentation

Benefits:
- Single responsibility: Each module handles one aspect
- Easier testing: NULL handling and analysis tested in isolation
- Better documentation: Each file documents its specific concerns
- Feature gating: #[cfg(feature = "simd")] guards work correctly

All 34 expression module tests pass. No performance regression.

Closes #2702
